refactor: split responsibilities on dto package
