+++ src/main/java/com/activegym/activegym/dto/memberships/MembershipTransferDTO.java
@@ -1,0 +1,4 @@
+package com.activegym.activegym.dto.memberships;
+
+public class MembershipTransferDTO {
+}
