+++ src/main/java/com/activegym/activegym/dto/users/UserDTO.java
@@ -1,4 +1,4 @@
-package com.activegym.activegym.dto;
+package com.activegym.activegym.dto.users;
 
 import lombok.Getter;
 import lombok.Setter;
