+++ src/main/java/com/activegym/activegym/dto/users/UserResponseDTO.java
@@ -1,30 +1,33 @@
-package com.activegym.activegym.dto;
+package com.activegym.activegym.dto.users;
 
 import lombok.Getter;
 import lombok.Setter;
 
 import java.time.LocalDate;
+import java.util.Set;
 
 /**
- * Data Transfer Object for creating an administrator.
- * This DTO includes the necessary information for administrator creation
- * and differs from {@link UserDTO} in that it requires a password for the new administrator account.
+ * Data Transfer Object for user response data.
+ * This DTO is used to represent the user information returned in API responses,
+ * including basic user details, the user's age and roles.
  *
  * @since v1.0
  * @author Carlos Esteban Castro Caicedo
  */
 @Getter
 @Setter
-public class AdminDTO {
+public class UserResponseDTO {
+    private Long id;
     private String document;
     private String firstName;
     private String lastName;
     private String phone;
     private String email;
-    private String password;
     private LocalDate dateOfBirth;
     private String genderName;
     private String epsName;
     private String bloodTypeName;
     private String bloodRhName;
+    private int age;
+    private Set<String> roles;
 }
