+++ src/main/java/com/activegym/activegym/service/Memberships/MembershipTypeService.java
@@ -1,7 +1,7 @@
 package com.activegym.activegym.service.Memberships;
 
 
-import com.activegym.activegym.dto.MembershipTypeDTO;
+import com.activegym.activegym.dto.memberships.MembershipTypeDTO;
 import com.activegym.activegym.exceptions.MembershipTypeNotFoundException;
 import com.activegym.activegym.model.Memberships.MembershipType;
 import com.activegym.activegym.repository.Memberships.MembershipTypeRepository;
